# v3/api/openapi.yaml
openapi: 3.1.0
info:
  title: Sakura v3 API
  version: 3.0.0-alpha.1
  description: Unified API used equally by the Web portal, Telegram Bot and API clients.
servers:
  - url: /
paths:
...
  /api/v3/admin/roles:
    get:
      operationId: listRoles
      tags: [RBAC]
      security: [{cookieAuth: []}]
      responses:
        "200": {description: Roles and permissions}
...
  /api/v3/admin/roles/{code}:
    parameters:
      - name: code
        in: path
        required: true
        schema: {type: string}
    put:
      operationId: saveCustomRole
      tags: [RBAC]
      security: [{cookieAuth: [], csrfHeader: []}]
      responses:
        "200": {description: Custom role saved}
        "403": {$ref: "#/components/responses/Forbidden"}
...
  /api/v3/admin/audit:
    get:
      operationId: listAuditLogs
      tags: [Audit]
      security: [{cookieAuth: []}]
      responses:
        "200": {description: Immutable management audit trail}
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sakura_v3_session
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
    bearerAuth:
      type: http
      scheme: bearer
...
  responses:
...
    Forbidden: {description: Permission, CSRF token, or scope denied}
...